+++ tests/Feature/resources/views/pages/podcasts/[.Tests.Feature.Fixtures.Podcast].blade.php
@@ -1,6 +1,9 @@
 <?php
 
 use function Laravel\Folio\middleware;
+use function Laravel\Folio\name;
+
+name('podcasts.show');
 
 middleware(function ($request, $next) {
     $_SERVER['__folio_podcasts_inline_middleware'] = $request->route('podcast');